send email request from CAMS
