+++ apps/contoso-fiber/src/main/java/com/contoso/cams/CamsApplication.java
@@ -16,6 +16,10 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 
+import com.azure.spring.messaging.checkpoint.Checkpointer;
+
+import static com.azure.spring.messaging.AzureHeaders.CHECKPOINTER;
+
 import nz.net.ultraq.thymeleaf.layoutdialect.LayoutDialect;
 
 @SpringBootApplication
@@ -46,9 +50,16 @@ public class CamsApplication {
     }
 
     @Bean
-    Consumer<Message<String>> consumeEmailResponse() {
+    Consumer<Message<String>> consumeemailresponse() {
         return message -> {
+            Checkpointer checkpointer = (Checkpointer) message.getHeaders().get(CHECKPOINTER);
             log.info("Received message: {}", message.getPayload());
+
+            // Checkpoint after processing the message
+            checkpointer.success()
+                .doOnSuccess(s -> log.info("Message '{}' successfully checkpointed", message.getPayload()))
+                .doOnError(e -> log.error("Exception found", e))
+                .block();
         };
     }
 }
