+++ apps/contoso-fiber/src/main/java/com/contoso/cams/config/SupportGuideSenderConfig.java
@@ -1,9 +1,14 @@
-package com.contoso.cams.support;
+package com.contoso.cams.config;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.cloud.stream.function.StreamBridge;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+
+import com.contoso.cams.services.SupportGuideDemoQueueSender;
+import com.contoso.cams.services.SupportGuideEmailSender;
+import com.contoso.cams.services.SupportGuideQueueSender;
+import com.contoso.cams.services.SupportGuideSender;
 
 @Configuration
 public class SupportGuideSenderConfig {
