+++ apps/contoso-fiber/src/main/java/com/contoso/cams/services/SupportGuideDemoQueueSender.java
@@ -1,4 +1,4 @@
-package com.contoso.cams.support;
+package com.contoso.cams.services;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -6,14 +6,14 @@ import org.springframework.cloud.stream.function.StreamBridge;
 
 import com.contoso.cams.protobuf.email.v1.EmailRequest;
 
-public class SupportGuideQueueSender implements SupportGuideSender {
+public class SupportGuideDemoQueueSender implements SupportGuideSender {
 
     private static final Logger log = LoggerFactory.getLogger(SupportGuideDemoQueueSender.class);
     private static final String EMAIL_REQUEST_QUEUE = "produceemailrequest-out-0";
 
     private final StreamBridge streamBridge;
 
-    public SupportGuideQueueSender(StreamBridge streamBridge) {
+    public SupportGuideDemoQueueSender(StreamBridge streamBridge) {
         this.streamBridge = streamBridge;
     }
 
@@ -27,7 +27,11 @@ public class SupportGuideQueueSender implements SupportGuideSender {
 
         log.info("EmailRequest: {}", emailRequest);
 
-        streamBridge.send(EMAIL_REQUEST_QUEUE, emailRequest);
+        var message = emailRequest.toString();
+
+        for (int i = 0; i < 1_000; i++) {
+            streamBridge.send(EMAIL_REQUEST_QUEUE, message);
+        }
 
         log.info("Message sent to the queue");
     }
