+++ apps/contoso-fiber/src/main/java/com/contoso/cams/services/SupportGuideQueueSender.java
@@ -1,4 +1,4 @@
-package com.contoso.cams.support;
+package com.contoso.cams.services;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -6,14 +6,14 @@ import org.springframework.cloud.stream.function.StreamBridge;
 
 import com.contoso.cams.protobuf.email.v1.EmailRequest;
 
-public class SupportGuideDemoQueueSender implements SupportGuideSender {
+public class SupportGuideQueueSender implements SupportGuideSender {
 
     private static final Logger log = LoggerFactory.getLogger(SupportGuideDemoQueueSender.class);
     private static final String EMAIL_REQUEST_QUEUE = "produceemailrequest-out-0";
 
     private final StreamBridge streamBridge;
 
-    public SupportGuideDemoQueueSender(StreamBridge streamBridge) {
+    public SupportGuideQueueSender(StreamBridge streamBridge) {
         this.streamBridge = streamBridge;
     }
 
@@ -27,12 +27,8 @@ public class SupportGuideDemoQueueSender implements SupportGuideSender {
 
         log.info("EmailRequest: {}", emailRequest);
 
-        //var message = emailRequest.toByteArray();
-
-        for (int i = 0; i < 1_000; i++) {
-            streamBridge.send(EMAIL_REQUEST_QUEUE, emailRequest);
-            //serviceBusSenderClient.sendMessage(new ServiceBusMessage(message));
-        }
+        var message = emailRequest.toString();
+        streamBridge.send(EMAIL_REQUEST_QUEUE, message);
 
         log.info("Message sent to the queue");
     }
