+++ apps/contoso-fiber/src/main/java/com/contoso/cams/services/SupportGuideSender.java
@@ -1,4 +1,4 @@
-package com.contoso.cams.support;
+package com.contoso.cams.services;
 
 public interface SupportGuideSender {
     void send(String to, String guideUrl, Long requestId);
