+++ apps/contoso-fiber/src/main/resources/application.properties
@@ -2,7 +2,7 @@
 contoso.retry.demo=0
 
 ## email, queue, demo
-contoso.suport-guide.request.service=email
+contoso.suport-guide.request.service=${CONTOSO_SUPPORT_GUIDE_REQUEST_SERVICE}
 
 # Spring Data JPA
 spring.datasource.url=${DATABASE_URL}
@@ -26,11 +26,13 @@ spring.data.redis.ssl.enabled=true
 spring.session.redis.namespace=spring:session
 
 # Spring Cloud Stream - Azure Service Bus
+spring.cloud.function.definition=consumeemailresponse
 spring.cloud.azure.servicebus.namespace=${AZURE_SERVICEBUS_NAMESPACE}
-spring.cloud.stream.bindings.produceemailrequest-out-0.destination=${AZURE_SERVICEBUS_EMAIL_RESPONSE_QUEUE_NAME}
 spring.cloud.stream.servicebus.bindings.produceemailrequest-out-0.producer.entity-type=queue
-spring.cloud.stream.bindings.consumeemailresponse-in-0.destination=${AZURE_SERVICEBUS_EMAIL_REQUEST_QUEUE_NAME}
-spring.cloud.stream.servicebus.bindings.consumeemailresponse-in-0.producer.entity-type=queue
+spring.cloud.stream.servicebus.bindings.consumeemailresponse-in-0.consumer.entity-type=queue
+spring.cloud.stream.servicebus.bindings.consumeemailresponse-in-0.consumer.auto-complete: false
+spring.cloud.stream.bindings.produceemailrequest-out-0.destination=${AZURE_SERVICEBUS_EMAIL_REQUEST_QUEUE_NAME}
+spring.cloud.stream.bindings.consumeemailresponse-in-0.destination=${AZURE_SERVICEBUS_EMAIL_RESPONSE_QUEUE_NAME}
 
 # Actuator
 management.endpoints.web.exposure.include=metrics,health,info,retry,retryevents
