+++ pom.xml
@@ -29,7 +29,7 @@
         <protobuf.version>3.25.3</protobuf.version>
         <protobuf.maven.plugin.version>0.6.1</protobuf.maven.plugin.version>
         <os.maven.plugin.version>1.7.1</os.maven.plugin.version>
-        <version.spring.boot>3.2.3</version.spring.boot>
+        <version.spring.boot>3.2.5</version.spring.boot>
         <commons-lang3.version>3.14.0</commons-lang3.version>
         <spring-cloud-azure.version>5.12.0</spring-cloud-azure.version>
 		<spring-cloud.version>2023.0.1</spring-cloud.version>
